pyproject.toml
[tool.poetry]
name = "newittest"
version = "0.1.0"
description = "test task"
authors = ["Fotlex <kudron02@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
pillow = "^12.0.0"
psycopg2-binary = "^2.9.11"
pydantic-settings = "^2.11.0"
django-cors-headers = "^4.9.0"
django-admin-rangefilter = "^0.13.3"
djangorestframework = "^3.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
